Refactored pack sub-packages
