+++ pkg/handlers/pipeline/scanner.go
