+++ pkg/packs/pipeline/commands/on.go
@@ -4,7 +4,7 @@ package commands
 import (
 	"github.com/arthur-debert/dodot/pkg/datastore"
 	"github.com/arthur-debert/dodot/pkg/filesystem"
-	handlerpipeline "github.com/arthur-debert/dodot/pkg/handler/pipeline"
+	handlerpipeline "github.com/arthur-debert/dodot/pkg/handlers/pipeline"
 	"github.com/arthur-debert/dodot/pkg/logging"
 	"github.com/arthur-debert/dodot/pkg/packs/commands"
 	"github.com/arthur-debert/dodot/pkg/packs/pipeline"
